# Improve ToString by 20%