@Fenny 2024-02-09
> The previous benchmark had 10x more performance. https://github.com/gofiber/utils/actions/runs/7753577848/job/21145172120
> 
> Is this related to the change in the benchmark code?

Yes, see https://github.com/gofiber/utils/pull/62#discussion_r1483825816